rename api.py to ensemble.py
