+++ libensemble/tests/functionality_tests/test_1d_sampling_from_files.py
@@ -17,7 +17,7 @@ The number of concurrent evaluations of the objective function will be 4-1=3.
 
 import numpy as np
 
-from libensemble.api import Ensemble
+from libensemble.ensemble import Ensemble
 
 # Main block is necessary only when using local comms with spawn start method (default on macOS and Windows).
 if __name__ == "__main__":
