+++ libensemble/tests/scaling_tests/forces/forces_adv/run_libe_forces_from_yaml.py
@@ -4,7 +4,7 @@ import sys
 
 import numpy as np
 
-from libensemble.api import Ensemble
+from libensemble.ensemble import Ensemble
 from libensemble.executors.mpi_executor import MPIExecutor
 
 ####################
