+++ libensemble/tests/scaling_tests/forces/funcx_forces/run_libe_forces_funcx.py
@@ -3,7 +3,7 @@ import secrets
 
 import numpy as np
 
-from libensemble.api import Ensemble
+from libensemble.ensemble import Ensemble
 
 if __name__ == "__main__":
     forces = Ensemble()
